create modal component, move the alert to components
