+++ src/assets/scss/_variables.scss
@@ -3,3 +3,5 @@ $purple-blue: #AAC4FF;
 $purple-blue-light: #D2DAFF;
 $purple-light: #EEF1FF;
 $blue: #37b7ff;
+$black: #000000;
+$white: #FFFFFF;
