+++ src/components/Alert/alert.scss
@@ -1,0 +1,8 @@
+.form-container {
+    align-items: stretch;
+    display: flex;
+    flex-direction: column;
+    input {
+        margin-bottom: 1rem;
+    }
+}
