+++ src/components/Modal/modal.scss
@@ -1,0 +1,47 @@
+@import "../../assets/scss//variables";
+
+.overlay {
+    align-items: center;
+    background-color: rgba(0, 0, 0, 0.5);
+    display: flex;
+    height: 100vh;
+    justify-content: center;
+    left: 0;
+    position: fixed;
+    top: 0;
+    width: 100vw;
+}
+.modal-container {
+    background-color: $white;
+    border-radius: 5px;
+    box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
+    min-height: 100px;
+    padding: 20px;
+    position: relative;
+    width: 80vw;
+    svg {
+        color: $purple-blue-light;
+        height: 20px;
+        position: absolute;
+        right: 20px;
+        top: 20px;
+        width: 20px;
+        &:hover {
+            color: $purple-blue;
+        }
+    }
+}
+.modal-header {
+    align-items: center;
+    border-bottom: 1px solid $purple-light;
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 20px;
+    padding-bottom: 10px;
+    h3 {
+        color: $blue;
+        font-size: 1rem;
+        font-weight: 500;
+        margin: 0;
+    }
+}
